requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0
gunicorn==20.0.4
